Renamed "filter" to "source-type-picker"
e.g. `<umb-property-editor-ui-tree-picker-source-type-picker>`
