+++ src/packages/core/property-editor/uis/tree-picker/config/source-type-picker/manifests.ts
@@ -2,11 +2,11 @@ import type { ManifestPropertyEditorUi } from '@umbraco-cms/backoffice/extension
 
 export const manifest: ManifestPropertyEditorUi = {
 	type: 'propertyEditorUi',
-	alias: 'Umb.PropertyEditorUi.TreePicker.Filter',
-	name: 'Tree Picker Filter Property Editor UI',
-	js: () => import('./property-editor-ui-tree-picker-filter.element.js'),
+	alias: 'Umb.PropertyEditorUi.TreePicker.SourceTypePicker',
+	name: 'Tree Picker Source Type Picker Property Editor UI',
+	js: () => import('./property-editor-ui-tree-picker-source-type-picker.element.js'),
 	meta: {
-		label: 'Tree Picker Filter',
+		label: 'Tree Picker Source Type Picker',
 		icon: 'icon-page-add',
 		group: 'pickers',
 		propertyEditorSchemaAlias: '',
